% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest neighbors cross-validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class of the training data.}

\item{k_nn}{Integer indicating that the model will use \code{k_nn} nearest neighbors.}

\item{k_cv}{Integer indicating that we will use \code{k_cv}-fold cross-validation.}
}
\value{
a list with class, a vector of the predicted class for each observation,
and cv_err, a numeric with the cross-validation misclassification error.
}
\description{
This function uses cross validation to assess how predictions made using
k-nearest neighbors models can generalize to independent data sets.
}
\examples{
train <- na.omit(my_penguins)[, c(3:6)]
cl <- na.omit(my_penguins)[, 1]
my_knn_cv(train, cl, 1, 5)

}
\keyword{prediction}
