% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{lm function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Class object.}

\item{data}{Input data frame.}
}
\value{
Table that includes the estimates for the regression coefficients,
the standard errors of the estimations, the t-values of the estimations, and
the probability of observing a t-value at least as extreme as the one observed,
if the null hypothesis is correct.
}
\description{
This function fits a linear model to given data.
}
\examples{
my_lm(formula = mtcars$wt ~ mtcars$cyl + mtcars$mpg, data = mtcars)

}
\keyword{inference}
\keyword{prediction}
