% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{String that specifies the alternative hypothesis and accepts
"two.sided", "less", or "greater".}

\item{mu}{Numeric value that specifices the null hypothesis.}
}
\value{
List that includes a numeric test statistic, the degrees of freedom,
the alternative hypothesis parameter, and the numeric p-value.
}
\description{
This function performs t-tests.
}
\examples{
my_t.test(x = rnorm(50, 0, 1), alternative = "less", mu = 0.3)
my_t.test(x = mtcars$wt, alternative = "greater", mu = 3)

}
\keyword{inference}
