% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random forest cross-validation
This function uses cross validation to assess how predictions made using
random forest models can generalize to independent data sets.}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric indicating that the function will use \code{k}-fold cross-validation.}
}
\value{
Numeric with the cross-validation error.
}
\description{
Random forest cross-validation
This function uses cross validation to assess how predictions made using
random forest models can generalize to independent data sets.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
